Capstone Projects/DataVisualiser/requirements.txt
altair==4.2.2
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dacite==1.8.1
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_colorscales==0.0.4
dash_daq==0.5.0
dtale==3.15.1
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Compress==1.17
fonttools==4.54.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
kiwisolver==1.4.7
llvmlite==0.43.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
missingno==0.5.2
multimethod==1.12
narwhals==1.13.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
plotly==5.24.1
protobuf==5.28.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.20.1
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
squarify==0.4.4
statsmodels==0.14.4
streamlit==1.39.0
strsimpy==0.2.1
sweetviz==2.3.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.6
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visions==0.7.6
watchdog==5.0.3
Werkzeug==3.0.6
wordcloud==1.9.3
xarray==2024.10.0
xlrd==2.0.1
ydata-profiling==4.12.0
zipp==3.20.2
zstandard==0.23.0

pandasgui~=0.2.14
autoviz~=0.1.905
lux~=0.5.1